[feat] 스터디 그룹 조회/ 생성 / 수정 / 삭제 api연결

+++ domain/src/main/java/com/umc/domain/model/request/organization/CreateStudyGroupRequest.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 data class CreateStudyGroupRequest(
     val name: String,
     val part: String,
-    val leaderId: Int,
-    val memberIds: List<Int>
+    val leaderId: Long,
+    val memberIds: List<Long>,
 )
